% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdccommdes.R
\name{cdccommdes}
\alias{cdccommdes}
\title{Community Design NCDM R Code}
\description{
R code for creating the following NCDMs developed by the CDC Community
Design workgroup circa 2018-2021:

\itemize{
\item \href{../doc/create_fars_indicators.html}{Traffic Safety}
\item \href{../doc/create_poverty_indicators.html}{Poverty}
\item Intersection Density
\item Population Density
\item Vehicle Miles Traveled
}
}
\details{
{
(Link other tutorials as completed directly in the list above.)

This package was made possible through funding from the CDC.
Its content is solely the responsibility of the authors and does not
necessarily represent the official views of the CDC.

}
}
\author{
Abigail Stamm, Bureau of Environmental and Occupational Epidemiology,
New York State Department of Health \cr
Margaret Horton, Colorado Department of Health \cr
}
