% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_intersections.R
\name{count_intersections}
\alias{count_intersections}
\title{Count road intersections}
\usage{
count_intersections(centroids, tractshp)
}
\arguments{
\item{centroids}{Road intersection point layer.}

\item{tractshp}{Census tract layer.}
}
\description{
Count the number of intersections in an area.
}
