% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_intersections.R
\name{create_intersections}
\alias{create_intersections}
\title{Create road intersections}
\usage{
create_intersections(path, shp)
}
\arguments{
\item{path}{File path.}

\item{shp}{Shapefile name.}
}
\description{
Create the road intersections.
}
