% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_fars.R
\name{join_fars}
\alias{join_fars}
\title{Join FARS accident locations to tract}
\usage{
join_fars(fars, tract)
}
\arguments{
\item{fars}{Dataset containing FARS data (data.frame or similar).}

\item{tract}{Simple feature containing tract information.}
}
\description{
This function joins the coordinates of FARS accidents to the tracts in
which they occur. This function is short, but takes several minutes to run
depending on the number of points in your dataset.
}
