% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_arc_aadt.R
\name{prep_arc_aadt}
\alias{prep_arc_aadt}
\title{Read in and prep ArcGIS Online street layer}
\usage{
prep_arc_aadt(my_state = "RI", year = 2019, path = NULL)
}
\arguments{
\item{my_state}{String denoting state abbreviation or FIPS code.}

\item{year}{String or number denoting data year.}

\item{path}{Path to save the resulting layer, if desired. Two layers are
saved: "clean_arc_streets.shp" contains all streets and
"aadt_arc_streets.shp" (the output object) contains only
streets with non-missing AADT values.}
}
\description{
This function pulls the HPMS layer of interest from the ArcGIS Online server,
isolates the variables of interest (AADT and street name), and converts it
from multilinestring to linestring to make processing easier.
}
