% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_water_areas.R
\name{prep_water_areas_obsolete}
\alias{prep_water_areas_obsolete}
\title{Prepare intersection density water areas}
\usage{
prep_water_areas_obsolete(shp_path, gdb, shp_name, save_path = NULL)
}
\arguments{
\item{shp_path}{String denoting absolute or relative path.}

\item{gdb}{String denoting the geodatabase name.}

\item{shp_name}{String denoting the geodatabase layer name.}

\item{save_path}{String denoting the path to save the processed
shapefile, if any.

Download your state's shapefile from
https://apps.nationalmap.gov/downloader/#/

The buffering step removes self-crossing boundaries, which throw errors
in later processing steps if not addressed.}
}
\description{
This function reads in the desired shapefile, subsets only developable
land, and buffers the polygons.
}
