% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_fars.R
\name{process_fars}
\alias{process_fars}
\title{Create the FARS indicators}
\usage{
process_fars(readpath, years, my_state, data_check = FALSE, tract_year = 2010)
}
\arguments{
\item{readpath}{String denoting your absolute or relative file path.}

\item{years}{Vector of years for which you have data.}

\item{my_state}{State FIPS code.}

\item{data_check}{Boolean that indicates whether or not R should produce a
simple map at the end that overlays accident locations on
census tracts.}

\item{tract_year}{Number denoting census tract vintage.

If you include the data check, plotting may be a bit slow.}
}
\description{
This function combines all FARS functions to create the full dataset of
FARS indicators. It outputs a final dataset for you to save in whatever
format you desire.
}
