% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fars.R
\name{read_fars}
\alias{read_fars}
\title{Read in FARS data}
\usage{
read_fars(filepath, years, my_state)
}
\arguments{
\item{filepath}{String denoting your absolute or relative file path.}

\item{years}{Vector of years for which you have data.}

\item{my_state}{State FIPS code.

To use this function, download the FARS zip files from
the \href{https://www.nhtsa.gov/node/97996/251}{FARS website}
and uzip all files to their default folders within the same folder.
Use the path to this folder as your filepath in the function. Remember,
R uses forward slash (/) for file paths. Make sure all years in your years
vector are in the folder (they need not be sequential) or R will throw a
missing file error.}
}
\description{
This function reads in and combines the Accident, Person, and PB type
datasets, which are the ones used for the traffic fatality indicators.
}
