% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fars_tract.R
\name{read_fars_tract}
\alias{read_fars_tract}
\title{Read census tract file}
\usage{
read_fars_tract(filepath, shp_name, id)
}
\arguments{
\item{filepath}{String denoting your absolute or relative file path.}

\item{shp_name}{String denoting the name of the shapefile.}

\item{id}{String denoting the variable name for the 11-digit tract
number.

Download your census tract file from the
\href{https://www.census.gov/cgi-bin/geo/shapefiles/index.php}{US Census}
and place it in the folder denoted in your filepath. Check which variable
contains the 11-digit tract number to place in the id option.}
}
\description{
This function reads in the census tract file and standardizes the name of
the 11-digit tract number for the indicator calculations.
}
