% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_land_areas.R
\name{read_land_areas}
\alias{read_land_areas}
\title{Prepare developable land layer}
\usage{
read_land_areas(shp_path, shp_name, save_path = NULL)
}
\arguments{
\item{shp_path}{String denoting absolute or relative path.}

\item{shp_name}{String denoting shapefile name.}

\item{save_path}{String denoting the path to save the processed
shapefile, if any.

Download your state's shapefile from
\href{https://www.usgs.gov/core-science-systems/science-analytics-and-synthesis/gap/science/pad-us-data-download?qt-science_center_objects=0#qt-science_center_objects}{USGS}

The subsetting step removes the following areas:
\itemize{
\item "Wild and Scenic River” (WSR)
\item “Special Designation Area” (SDA)
\item “National Scenic, Botanical or Volcanic” (NSBV)
}

The buffering step removes self-crossing boundaries, which throw errors
in later processing steps if not addressed.}
}
\description{
This function reads in the desired shapefile, subsets only developable
land, and buffers the polygons.
}
