% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_streets.R
\name{read_streets}
\alias{read_streets}
\title{Read streets shapefile}
\usage{
read_streets(path, shp)
}
\arguments{
\item{path}{File path.}

\item{shp}{Streets shapefile.}
}
\description{
This function reads in the streets shapefile, isolates only the street types
of interest, combines smaller segments of the same street, and returns a
clean street layer.
}
