% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tract.R
\name{read_tract}
\alias{read_tract}
\title{Read census tract file}
\usage{
read_tract(my_state, year = 2010)
}
\arguments{
\item{my_state}{String denoting state abbreviation or FIPS code.}

\item{year}{String or number denoting data year.}
}
\description{
This function reads in the census tract file from the tigris package and
standardizes the name of the 11-digit tract number for the indicator
calculations.
}
