% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_fars.R
\name{summarize_fars}
\alias{summarize_fars}
\title{Summarize FARS accidents}
\usage{
summarize_fars(fars, tract, my_state)
}
\arguments{
\item{fars}{Simple feature containing FARS coordinates linked to
census tracts.}

\item{tract}{Simple feature containing census tracts.}

\item{my_state}{State FIPS code.}
}
\description{
This function summarizes deaths and accidents by year and census tract.
}
