% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_buffer_centroid.r
\name{calc_buffer_centroid}
\alias{calc_buffer_centroid}
\title{Calculate buffer centroid}
\usage{
calc_buffer_centroid(area, width = 30, dissolve = TRUE)
}
\arguments{
\item{area}{Point layer.}

\item{width}{Buffer radius.}

\item{dissolve}{Whether to merge overlapping buffers,
but this option is slow.}
}
\description{
Calculate the centroid for a set of very close intersections
(within a defined buffer).
}
