% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_intersections.R
\name{create_intersections}
\alias{create_intersections}
\title{Create road intersections}
\usage{
create_intersections(shp, save_path = NULL)
}
\arguments{
\item{shp}{Cleaned streets layer.}

\item{save_path}{String denoting the path to save the processed
shapefile, if any.}
}
\description{
Create the road intersections. This function takes a long time to run due
to the massive size of road shapefiles.
}
