% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_fars.R
\name{process_fars}
\alias{process_fars}
\title{Create the FARS indicators}
\usage{
process_fars(readpath, shp_name, years, tract_id, my_state, data_check = FALSE)
}
\arguments{
\item{readpath}{String denoting your absolute or relative file path.}

\item{shp_name}{String denoting the name of the shapefile.}

\item{years}{Vector of years for which you have data.}

\item{tract_id}{String denoting the variable name for the 11-digit tract
number.}

\item{my_state}{State FIPS code.}

\item{data_check}{Boolean that indicates whether or not R should produce a
simple map at the end that overlays accident locations on
census tracts.

If you include the data check, plotting may be a bit slow.}
}
\description{
This function combines all FARS functions to create the full dataset of
FARS indicators. It outputs a final dataset for you to save in whatever
format you desire.
}
