% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_streets.R
\name{read_streets}
\alias{read_streets}
\title{Read streets shapefile}
\usage{
read_streets(shp_path, shp_name, save_path = NULL)
}
\arguments{
\item{shp_path}{String denoting absolute or relative path, which may
include the geodatabase name, if relevant.}

\item{shp_name}{String denoting the shapefile or geodatabase layer name.}

\item{save_path}{String denoting the path to save the processed
shapefile, if any.}
}
\description{
This function reads in the streets shapefile, isolates only the street types
of interest, combines smaller segments of the same street, and returns a
clean street layer.
}
